refactoring for the createSurvey page
